requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    poetry.lock --requirements
#

django = "^5.0.1"
python-dotenv = "^1.0.0"
pillow = "^10.2.0"